+++ conformance/interfaces/mmap/14-1.c
@@ -68,7 +68,7 @@ int main()
   
   data = (char *) malloc(total_size); 
   memset(data, 'a', total_size);
-  printf("Time before write(): %ld\n", time(NULL));
+  printf("Time before write(): %ld\n", (long int)time(NULL));
   if (write(fd, data, total_size) != total_size)
   {
     printf(TNAME "Error at write(): %s\n", 
@@ -80,7 +80,7 @@ int main()
   sleep(1); 
   flag = MAP_SHARED;
   prot = PROT_READ | PROT_WRITE;
-  printf("Time before mmap(): %ld\n", time(NULL));
+  printf("Time before mmap(): %ld\n", (long int)time(NULL));
   pa = mmap(addr, size, prot, flag, fd, off);
   if (pa == MAP_FAILED)
   {
@@ -90,7 +90,7 @@ int main()
     exit(PTS_FAIL);
   }
   sleep(1); 
-  printf("Time before write reference: %ld\n", time(NULL));
+  printf("Time before write reference: %ld\n", (long int)time(NULL));
   /* Before write reference */
   if (stat(tmpfname, &stat_buff) == -1)
   {
@@ -110,7 +110,7 @@ int main()
   * is not acurate enough to reflect the update
   */
   sleep(1);
-  printf("Time before msync(): %ld\n", time(NULL));
+  printf("Time before msync(): %ld\n", (long int)time(NULL));
   msync(pa, size, MS_SYNC);
 
   /* FIXME: Update the in-core meta data to the disk */
@@ -128,7 +128,7 @@ int main()
   mtime2 = stat_buff.st_mtime;
  
   printf("ctime1: %ld, ctime2: %ld\nmtime1: %ld, mtime2: %ld\n",
-                  ctime1, ctime2, mtime1, mtime2); 
+                  (long int)ctime1, (long int)ctime2, (long int)mtime1, (long int)mtime2); 
   if (ctime2 == ctime1 || mtime2 == mtime1)
   {
     printf("Test Fail " TNAME
